+++ tiled/catalog/adapter.py
@@ -72,7 +72,7 @@ from ..mimetypes import (
 )
 from ..query_registration import QueryTranslationRegistry
 from ..server.core import NoEntry
-from ..server.schemas import Asset, DataSource, Management, Revision, Spec
+from ..server.schemas import Asset, DataSource, Management, Revision
 from ..storage import (
     FileStorage,
     SQLStorage,
@@ -80,7 +80,7 @@ from ..storage import (
     parse_storage,
     register_storage,
 )
-from ..structures.core import StructureFamily
+from ..structures.core import Spec, StructureFamily
 from ..utils import (
     UNCHANGED,
     Conflicts,
@@ -156,7 +156,7 @@ class RootNode:
         self.id = 0
         self.parent = None
         self.metadata_ = metadata or {}
-        self.specs = [Spec.model_validate(spec) for spec in specs or []]
+        self.specs = specs or []
         self.key = ""
         self.data_sources = None
         self.access_blob = top_level_access_blob or {}
@@ -243,7 +243,7 @@ class CatalogNodeAdapter:
         self.conditions = conditions or []
         self.queries = queries or []
         self.structure_family = node.structure_family
-        self.specs = [Spec.model_validate(spec) for spec in node.specs]
+        self.specs = [Spec(**spec) for spec in node.specs]
         self.startup_tasks = [self.startup]
         if mount_path:
             self.startup_tasks.append(partial(self.create_mount, mount_path))
@@ -291,8 +291,12 @@ class CatalogNodeAdapter:
     def writable(self):
         return bool(self.context.writable_storage)
 
+    @property
+    def key(self):
+        return self.node.key
+
     def __repr__(self):
-        return f"<{type(self).__name__} {self.node.key}>"
+        return f"<{type(self).__name__} {self.key}>"
 
     async def __aiter__(self):
         statement = select(orm.Node.key).filter(orm.Node.parent == self.node.id)
@@ -401,13 +405,6 @@ class CatalogNodeAdapter:
                         if adapter is None:
                             raise NoEntry(segments)
                     return adapter
-                elif (
-                    isinstance(catalog_adapter, CatalogCompositeAdapter)
-                    and len(segments[i:]) == 1
-                ):
-                    # Trying to access a table column or an array in a composite node
-                    _segm = await catalog_adapter.resolve_flat_key(segments[i])
-                    return await catalog_adapter.lookup_adapter(_segm.split("/"))
             raise NoEntry(segments)
 
         return STRUCTURES[node.structure_family](self.context, node)
@@ -578,7 +575,7 @@ class CatalogNodeAdapter:
             parent=self.node.id,
             metadata_=metadata,
             structure_family=structure_family,
-            specs=[s.model_dump() for s in specs or []],
+            specs=specs or [],
             access_blob=access_blob,
         )
         async with self.context.session() as db:
@@ -597,10 +594,7 @@ class CatalogNodeAdapter:
             await db.refresh(node)
             for data_source in data_sources:
                 if data_source.management != Management.external:
-                    if structure_family in {
-                        StructureFamily.container,
-                        StructureFamily.composite,
-                    }:
+                    if structure_family == StructureFamily.container:
                         raise NotImplementedError(structure_family)
                     if data_source.mimetype is None:
                         data_source.mimetype = DEFAULT_CREATION_MIMETYPE[
@@ -696,7 +690,7 @@ class CatalogNodeAdapter:
                     )
                 )
             ).scalar()
-            return key, type(self)(self.context, refreshed_node)
+            return type(self)(self.context, refreshed_node)
 
     async def _put_asset(self, db: AsyncSession, asset):
         # Find an asset_id if it exists, otherwise create a new one
@@ -911,7 +905,7 @@ class CatalogNodeAdapter:
             # SQLAlchemy reserved word 'metadata'.
             values["metadata_"] = metadata
         if specs is not None:
-            values["specs"] = [s.model_dump() for s in specs]
+            values["specs"] = specs
         if access_blob is not None:
             values["access_blob"] = access_blob
         async with self.context.session() as db:
@@ -1006,69 +1000,6 @@ class CatalogContainerAdapter(CatalogNodeAdapter):
                 return self.search(KeysFilter(fields))
             return self
         return await ensure_awaitable((await self.get_adapter()).read, *args, **kwargs)
-
-
-class CatalogCompositeAdapter(CatalogContainerAdapter):
-    async def resolve_flat_key(self, key):
-        for _key, item in await self.items_range(offset=0, limit=None):
-            if key == _key:
-                return _key
-            if item.structure_family == StructureFamily.table:
-                if key in item.structure().columns:
-                    return f"{_key}/{key}"
-        raise KeyError(key)
-
-    async def create_node(
-        self,
-        structure_family,
-        metadata,
-        key=None,
-        specs=None,
-        data_sources=None,
-        access_blob=None,
-    ):
-        key = key or self.context.key_maker()
-
-        # List all new keys that would be added to the flattened namespace
-        assert len(data_sources) == 1
-        if data_sources[0].structure_family == StructureFamily.table:
-            new_keys = data_sources[0].structure.columns
-        elif data_sources[0].structure_family in {
-            StructureFamily.array,
-            StructureFamily.awkward,
-            StructureFamily.sparse,
-        }:
-            new_keys = [key]
-        else:
-            raise ValueError(
-                f"Unsupported structure family: {data_sources[0].structure_family}"
-            )
-
-        # Get all keys and columns names already in the Composite node
-        flat_keys = []
-        for _key, item in await self.items_range(offset=0, limit=None):
-            flat_keys.append(_key)
-            if item.structure_family == StructureFamily.table:
-                flat_keys.extend(item.structure().columns)
-
-        # Check for column name collisions
-        key_conflicts = set(new_keys) & set(flat_keys)
-        if key_conflicts:
-            raise Collision(f"Column name collision: {key_conflicts}")
-
-        # Ensure that child tables are marked with the 'flattened' spec
-        if structure_family == StructureFamily.table:
-            if not specs or "flattened" not in (s.name for s in specs):
-                specs = [Spec(name="flattened")] + (specs or [])
-
-        return await super().create_node(
-            structure_family,
-            metadata,
-            key=key,
-            specs=specs,
-            data_sources=data_sources,
-            access_blob=access_blob,
-        )
 
 
 class CatalogArrayAdapter(CatalogNodeAdapter):
@@ -1671,7 +1602,6 @@ def node_from_segments(segments, root_id=0):
 STRUCTURES = {
     StructureFamily.array: CatalogArrayAdapter,
     StructureFamily.awkward: CatalogAwkwardAdapter,
-    StructureFamily.composite: CatalogCompositeAdapter,
     StructureFamily.container: CatalogContainerAdapter,
     StructureFamily.sparse: CatalogSparseAdapter,
     StructureFamily.table: CatalogTableAdapter,